+++ pagure/hooks/pagure_hook.py
@@ -94,7 +94,9 @@ class PagureHook(BaseHook):
             # Install the hook itself
             hook_path = os.path.join(
                 repopath, 'hooks', 'post-receive.pagure')
-            os.symlink(hook_file, hook_path)
+            hook_file = os.path.join(hook_files, 'git_multimail.py')
+            if not os.path.exists(hook_file):
+                os.symlink(hook_file, hook_path)
 
     @classmethod
     def remove(cls, project):
